Добавлены тесты по тестовым данным из файлов .md

+++ main.py
@@ -1,14 +1,14 @@
 def main_menu(score):
-    if 0 < int(score) <= 100:
+    if 0 < float(score) <= 100:
         percent = 1
         # print('Ваша скидка составляет ', percent, '%')
-    elif 100 < int(score) <= 200:
+    elif 100 < float(score) <= 200:
         percent = 3
         # print('Ваша скидка составляет ', percent, '%')
-    elif 200 < int(score) <= 500:
+    elif 200 < float(score) <= 500:
         percent = 5
         # print('Ваша скидка составляет ', percent, '%')
-    elif 500 < int(score):
+    elif 500 < float(score):
         percent = 10
         # print('Ваша скидка составляет ', percent, '%')
     return percent
